requirements.txt
flask
mysql-connector-python
pandas
matplotlib
scikit-learn
plotly
requests

